# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Digital Signage API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List devices, optionally filtered by client
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Create a device linked to a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          description: Bad request
        "409":
          description: Device key already linked
  /devices/by-key/{key}:
    get:
      operationId: getDeviceByKey
      tags: [devices]
      summary: Find the device that owns a device key (TV pairing)
      parameters:
        - { name: key, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          description: Invalid key format
        "404":
          description: Not found
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Not found
    patch:
      operationId: updateDevice
      tags: [devices]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Not found
    delete:
      operationId: deleteDevice
      tags: [devices]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /devices/{id}/playlist:
    get:
      operationId: getDevicePlaylist
      tags: [devices]
      summary: Get assigned playlist for a device
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlaylistItem"
        "404":
          description: Not found
  /devices/{id}/preview:
    get:
      operationId: getDevicePreview
      tags: [devices]
      summary: Slides the TV is showing right now, for the admin preview
      description: >
        Mesma rotação que /display/{deviceKey}/slides entrega à TV, com a
        origem de cada slide. Sem efeito colateral: não mexe em lastSeenAt.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DevicePreviewSlide"
        "404":
          description: Not found
...
  /devices/{id}/playlist/add:
    post:
      operationId: addToDevicePlaylist
      tags: [devices]
      summary: Assign an announcement to a device
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistItem"
        "400":
          description: Bad request
  /devices/{id}/playlist/reorder:
    post:
      operationId: reorderDevicePlaylist
      tags: [devices]
      summary: Reorder device playlist
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaylistReorder"
      responses:
        "200":
          description: OK
  /devices/{deviceId}/playlist/{announcementId}:
    delete:
      operationId: removeFromDevicePlaylist
      tags: [devices]
      summary: Remove an announcement from a device playlist
      parameters:
        - { name: deviceId, in: path, required: true, schema: { type: integer } }
        - { name: announcementId, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Removed
        "404":
          description: Not found
  /devices/{id}/playlist/{announcementId}/toggle:
    patch:
      operationId: togglePlaylistItem
      tags: [devices]
      summary: Toggle active status of a playlist item
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: announcementId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistItem"
        "404":
          description: Not found

  # ── Display (TV endpoint) ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Device:
      type: object
      required: [id, clientId, clientName, name, deviceKey, orientation, showcase, createdAt]
      properties:
        id: { type: integer }
        clientId: { type: integer }
        clientName: { type: string }
        name: { type: string }
        location: { type: ["string", "null"] }
        # "landscape" | "portrait_right" | "portrait_left". Sem enum pelo mesmo
        # motivo de Announcement.orientation: uma linha estranha no banco não
        # pode derrubar o parse da lista de TVs inteira.
        orientation: { type: string }
        deviceKey: { type: string }
        lastSeenAt: { type: ["string", "null"], format: date-time }
        showcase: { type: boolean }
        # Link do YouTube que toca em fundo na TV. Nulo = sem música.
        musicUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    DeviceInput:
      type: object
      required: [clientId, name]
      properties:
        clientId: { type: integer }
        name: { type: string, minLength: 1 }
        location: { type: string }
        deviceKey: { type: string, pattern: "^[0-9A-F]{16}$" }
    DeviceUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        location: { type: ["string", "null"] }
        orientation: { type: string, enum: [landscape, portrait_right, portrait_left] }
        showcase: { type: boolean }
        # Link de vídeo ou playlist do YouTube. null ou vazio tira a música.
        musicUrl: { type: ["string", "null"] }

    # ── Parque de TVs ────────────────────────────────────────────────────────
...
    PlaylistItem:
      type: object
      required: [id, deviceId, announcementId, displayOrder, isActive, title, imageUrl, duration, orientation]
      properties:
        id: { type: integer }
        deviceId: { type: integer }
        announcementId: { type: integer }
        displayOrder: { type: integer }
        isActive: { type: boolean }
        title: { type: string }
        # Nulo é estado normal da peça: nasce sem imagem quando não há upload, e
        # o poster é zerado ao trocar para YouTube. Proibir nulo aqui derrubava
        # as três rotas de playlist depois que a linha já estava no banco.
        imageUrl: { type: string, nullable: true }
        duration: { type: integer }
        # "landscape" | "portrait" da peça. O portal usa para avisar, na
        # playlist, o item que não toca na orientação atual da TV.
        orientation: { type: string }
    PlaylistItemInput:
      type: object
      required: [announcementId]
      properties:
        announcementId: { type: integer }
        displayOrder: { type: integer }
    PlaylistReorder:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items: { type: integer }

    # ── Display ──────────────────────────────────────────────────────────────
    DisplaySlide:
      type: object
      required: [announcementId, title, duration, mediaKind]
      properties:
        announcementId: { type: integer }
        campaignId: { type: integer, nullable: true }
        # `title` segue no payload só para não quebrar TVs com tv.html em cache
        # antigo. Quem vai ao ar é `displayText`, já resolvido pelo servidor:
        # null significa slide sem texto.
        title: { type: string }
        displayText: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }
        duration: { type: integer }
        qrImageUrl: { type: string, nullable: true }
        mediaKind: { type: string }
        youtubeId: { type: string, nullable: true }
        playbackMode: { type: string, nullable: true }
        audioMode: { type: string, nullable: true }
        videoIds:
          type: array
          nullable: true
          items: { type: string }
    DevicePreviewSlide:
      allOf:
        - $ref: "#/components/schemas/DisplaySlide"
        - type: object
          required: [source]
          properties:
            source:
              type: string
              enum: [campaign, panel, playlist, alert]
...